select items: refactor get_single_param and get_per_channel_param
These are no longer HOA-specific.
